requirements.txt
# Core dependencies
openai>=1.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0
aiofiles>=23.0.0

# CLI dependencies
colorama>=0.4.0

# API dependencies
flask>=3.0.0
flask-cors>=4.0.0
flask-socketio>=5.3.0
python-socketio>=5.10.0

# Document processing
python-docx>=1.0.0
docx2pdf>=0.1.8

# Utilities
requests>=2.31.0
pytz>=2023.3

# Azure support (optional)
# azure-storage-blob
# azure-servicebus

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Development
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0